Split assessment column in its own component
