+++ src/pages/application-inventory/components/application-assessment/index.ts
@@ -1,0 +1,1 @@
+export { ApplicationAssessment } from "./application-assessment";
